# api_spec.yml
openapi: 3.0.0
info:
  title: API Tienda de Productos - ABP
  version: 1.0.0
  description: EspecificaciÃ³n completa de rutas para el MVP de la tienda de productos.
paths:
  /api/auth/login:
    post:
      tags:
        - AutenticaciÃ³n
      summary: Iniciar sesiÃ³n de usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: AutenticaciÃ³n exitosa y retorno de sesiÃ³n/token
        '401':
          description: Credenciales incorrectas
 
  /api/usuarios:
    get:
      tags:
        - Usuarios
      summary: Obtener lista completa de usuarios
      responses:
        '200':
          description: Lista de usuarios recuperada
    post:
      tags:
        - AutenticaciÃ³n
      summary: Registrar un nuevo usuario del sistema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                password:
                  type: string
                nombre:
                  type: string
                apellido:
                  type: string
                dni:
                  type: string
                direccion:
                  type: string
                telefono:
                  type: string
                poblacion:
                  type: string
                provincia:
                  type: string
                codigo_postal:
                  type: integer
                rol_id:
                  type: integer
      responses:
        '201':
          description: Usuario registrado correctamente
        '400':
          description: Error en la validaciÃ³n de los datos entregados
 
  /api/usuarios/{id}:
    get:
      tags:
        - Usuarios
      summary: Obtener los detalles de un usuario por su ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Datos del usuario encontrados
        '404':
          description: El usuario solicitado no existe
 
  /api/productos:
    get:
      tags:
        - Productos
      summary: Listar todo el catÃ¡logo de productos
      responses:
        '200':
          description: Lista de productos obtenida
    post:
      tags:
        - Productos
      summary: Crear un nuevo producto en el catÃ¡logo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                properties:
                nombre:
                  type: string
                descripcion:
                  type: string
                precio:
                  type: number
                stock:
                  type: integer
                categoria_id:
                  type: integer
                id_tipo_producto:
                  type: integer
                url_producto:
                  type: string
                  format: uri
      responses:
        '201':
          description: Producto aÃ±adido con Ã©xito
        '400':
          description: ParÃ¡metros obligatorios faltantes
 
  /api/productos/{id}:
    get:
      tags:
        - Productos
      summary: Obtener la informaciÃ³n de un producto especÃ­fico
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Producto encontrado
        '404':
          description: Producto no encontrado
    put:
      tags:
        - Productos
      summary: Actualizar la informaciÃ³n o stock de un producto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                properties:
                nombre:
                  type: string
                descripcion:
                  type: string
                precio:
                  type: number
                stock:
                  type: integer
                categoria_id:
                  type: integer
                id_tipo_producto:
                  type: integer
                url_producto:
                  type: string
                  format: uri
      responses:
        '200':
          description: Producto actualizado correctamente
        '404':
          description: El producto a modificar no existe
    delete:
      tags:
        - Productos
      summary: Eliminar un producto del sistema
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Producto eliminado con Ã©xito
        '404':
          description: El producto no existe
 
  /api/deportes:
    get:
      tags:
        - Deportes
      summary: Listar todos los deportes disponibles
      responses:
        '200':
          description: Lista de deportes devuelta
    post:
      tags:
        - Deportes
      summary: Crear un nuevo deporte para la tienda
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
      responses:
        '201':
          description: Deporte creado con Ã©xito
 
  /api/tipos_producto:
    get:
      tags:
        - Tipos de Producto
      summary: Listar todos los tipos de producto disponibles
      responses:
        '200':
          description: Lista de tipos de producto devuelta
    post:
      tags:
        - Tipos de Producto
      summary: Crear un nuevo tipo de producto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
      responses:
        '201':
          description: Tipo de producto creado con Ã©xito
 
  /api/pedidos:
    get:
      tags:
        - Pedidos
      summary: Obtener el histÃ³rico de todos los pedidos realizados
      responses:
        '200':
          description: Lista de pedidos devuelta
    post:
      tags:
        - Pedidos
      summary: Tramitar un nuevo pedido con mÃºltiples productos vinculados
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                usuario_id:
                  type: integer
                direccion_envio:
                  type: string
                poblacion_envio:
                  type: string
                provincia_envio:
                  type: string
                codigo_postal_envio:
                  type: string
                estado_pedido:
                  type: boolean
                lineas:
                  type: array
                  items:
                    type: object
                    properties:
                      producto_id:
                        type: integer
                      cantidad:
                        type: integer
                      precio_unitario:
                        type: number
      responses:
        '201':
          description: Pedido y lÃ­neas de pedido creados correctamente
        '400':
          description: Error de stock o datos de envÃ­o errÃ³neos
 
  /api/pedidos/{id}:
    get:
      tags:
        - Pedidos
      summary: Consultar los detalles de un pedido y sus respectivas lÃ­neas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: InformaciÃ³n del pedido y sus productos asociada
        '404':
          description: El pedido consultado no existe
    put:
      tags:
        - Pedidos
      summary: Actualizar el estado de un pedido especÃ­fico
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                estado_pedido:
                  type: boolean
      responses:
        '200':
          description: Estado del pedido actualizado correctamente
        '404':
          description: El pedido solicitado no existe
 
  /api/roles:
    get:
      tags:
        - Roles
      summary: Obtener lista completa de roles
      responses:
        '200':
          description: Lista de roles recuperada
    post:
      tags:
        - Roles
      summary: Crear un nuevo rol
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
      responses:
        '201':
          description: Rol creado correctamente
 
  /api/roles/{id}:
    get:
      tags:
        - Roles
      summary: Obtener los detalles de un rol por su ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Datos del rol encontrados
        '404':
          description: El rol solicitado no existe
    delete:
      tags:
        - Roles
      summary: Eliminar un rol del sistema
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Rol eliminado con Ã©xito
        '404':
          description: El rol no existe